"LE.UM.4.2.2.r1-02800-SDX55"

* tag 'LE.UM.4.2.2.r1-02800-SDX55': (86 commits)
  sched: Fix out of bounds issue in for_each_cluster macro
  cnss2: Add cnss_smmu_unmap API
  net: stmmac: Fixing KW errors
  msm: ipa: disable status on ODL_DPL_CONS pipe above IPAv4.5 targets
  net: aquantia: Using non-interrupt mode for ICMP packets
  msm: mhi_dev: add WQ_UNBOUND and set max_active for mhi_sm_wq
  atlantic forwarding driver v1.1.9
  atlantic forwarding driver v1.1.8
  msm: camera: csiphy: Reset phy config count in phy release
  msm: camera: icp: lock before updating clock status
  msm: camera: sensor: unregister subdev if cpas registration fails
  msm: camera: jpeg: Reset all blocks during jpeg reset
  arm64: dts: QCS405: Cpu dai for SEN TDM interface
  msm: mhi_dev: Change to not enable ERDB interrupts
  msm: npu_v2: Continue npu_probe() when !DEBUG_FS
  coresight: tmc-etr: fix etr smmu unmap issue
  diag: Do not fail pcie cb registrations for EEXIST return
  media: uvcvideo: Avoid cyclic entity chains due to malformed USB descriptors
  mhi: cntrl: qcom: remove multiple calls to sysfs_create_group()
  msm: ipa: add new unit test case for wdi 2.4G Tx pipe
  ...


Conflicts:
	arch/arm64/boot/dts/qcom/Makefile
	drivers/char/diag/diag_pcie.c
	drivers/esoc/esoc-mdm-4x.c
	drivers/media/platform/msm/camera/cam_icp/icp_hw/bps_hw/bps_core.c
	drivers/media/platform/msm/camera/cam_icp/icp_hw/ipe_hw/ipe_core.c
	drivers/media/platform/msm/camera/cam_jpeg/jpeg_hw/jpeg_enc_hw/cam_jpeg_enc_hw_info_ver_4_2_0.h
	drivers/net/ethernet/aquantia/atlantic-fwd/atl_ring.c
	drivers/net/wireless/ath/ath10k/mac.c
	drivers/net/wireless/ath/ath10k/wmi-tlv.c
	drivers/platform/msm/mhi_dev/mhi_sm.c
	kernel/sched/core_ctl.c

+++ drivers/char/diag/diag_pcie.c
@@ -490,12 +490,9 @@ static void diag_pcie_connect(struct diag_pcie_info *ch)
 	queue_work(ch->wq, &(ch->read_work));
 }
 
-void diag_pcie_open_work_fn(struct work_struct *work)
+static void diag_pcie_open_channels(struct diag_pcie_info *pcie_info)
 {
 	int rc = 0;
-	struct diag_pcie_info *pcie_info = container_of(work,
-						      struct diag_pcie_info,
-						      open_work);
 
 	if (!pcie_info || atomic_read(&pcie_info->enabled))
 		return;
@@ -538,6 +535,15 @@ handle_in_err:
 handle_not_rdy_err:
 	mutex_unlock(&pcie_info->in_chan_lock);
 	mutex_unlock(&pcie_info->out_chan_lock);
+}
+
+void diag_pcie_open_work_fn(struct work_struct *work)
+{
+	struct diag_pcie_info *pcie_info = container_of(work,
+						      struct diag_pcie_info,
+						      open_work);
+
+	diag_pcie_open_channels(pcie_info);
 }
 
 /*
@@ -698,7 +704,10 @@ int diag_pcie_register(int id, int ctxt, struct diag_mux_ops *ops)
 	mutex_init(&ch->in_chan_lock);
 	mutex_init(&ch->out_chan_lock);
 	rc = diag_register_pcie_channels(ch);
-	if (rc < 0) {
+	if (rc == -EEXIST) {
+		diag_pcie_open_channels(ch);
+		pr_err("diag: Handled -EEXIST error\n");
+	} else if (rc < 0 && rc != -EEXIST) {
 		if (ch->wq)
 			destroy_workqueue(ch->wq);
 		kfree(ch->in_chan_attr.read_buffer);
